Extract ShiftedReadPlanner from MongoDbPlanner into separate file
